cijferlijst af

+++ ELO/Grade.cs
@@ -10,11 +10,11 @@ namespace ELO
 
         public Class _class { get; private set; }
 
-        public decimal grade { get; private set; }
+        public double grade { get; private set; }
         public string date { get; private set; }
         public string subject { get; private set; }
         public decimal weight { get; private set; }
-        public Grade(Student student, Class _class, decimal grade, string date, string subject, decimal weight)
+        public Grade(Student student, Class _class, double grade, string date, string subject, decimal weight)
         {
             this.student = student;
             this._class = _class;
